[COMMIT][CHAGE] Alterado locais dos arquivos e adicioando outro exemplo de aplicação do padrao factory
