+++ src/main/java/br/com/lima/creationalPatterns/factory/db/connections/ConnectionFactory.java
@@ -1,0 +1,8 @@
+package br.com.lima.creationalPatterns.factory.db.connections;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public interface ConnectionFactory {
+    Connection createConnection() throws SQLException;
+}
