+++ src/main/java/br/com/lima/creationalPatterns/factory/db/connections/MySqlConnectionFactory.java
@@ -1,0 +1,24 @@
+package br.com.lima.creationalPatterns.factory.db.connections;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class MySqlConnectionFactory implements ConnectionFactory {
+
+    private String url;
+    private String username;
+    private String password;
+
+
+    public MySqlConnectionFactory(String url, String username, String password) {
+        this.url = url;
+        this.username = username;
+        this.password = password;
+    }
+
+    @Override
+    public Connection createConnection() throws SQLException {
+        return DriverManager.getConnection(url, username, password);
+    }
+}
