--- src/main/java/br/com/lima/creationalPatterns/factory/factoryMethod/factory/CircleFactory.java
@@ -1,14 +1,0 @@
-package br.com.lima.creationalPatterns.factory.factoryMethod.factory;
-
-
-import br.com.lima.creationalPatterns.factory.factoryMethod.model.Circle;
-import br.com.lima.creationalPatterns.factory.factoryMethod.model.Shape;
-
-public class CircleFactory extends ShapeFactory{
-
-
-    @Override
-    protected Shape createShape() {
-        return new Circle();
-    }
-}
