--- src/main/java/br/com/lima/creationalPatterns/factory/factoryMethod/factory/RectangleFactory.java
@@ -1,13 +1,0 @@
-package br.com.lima.creationalPatterns.factory.factoryMethod.factory;
-
-import br.com.lima.creationalPatterns.factory.factoryMethod.model.Rectangle;
-import br.com.lima.creationalPatterns.factory.factoryMethod.model.Shape;
-
-public class RectangleFactory extends ShapeFactory{
-
-
-    @Override
-    protected Shape createShape() {
-        return new Rectangle();
-    }
-}
