--- src/main/java/br/com/lima/creationalPatterns/factory/factoryMethod/factory/ShapeFactory.java
@@ -1,18 +1,0 @@
-package br.com.lima.creationalPatterns.factory.factoryMethod.factory;
-
-
-import br.com.lima.creationalPatterns.factory.factoryMethod.model.Shape;
-
-public abstract class ShapeFactory {
-
-    public Shape orderShape() {
-        Shape shape = createShape();
-
-        shape.calculateArea();
-
-        return shape;
-    }
-
-    protected abstract Shape createShape();
-
-}
