--- src/main/java/br/com/lima/creationalPatterns/factory/halfSimple/enumeration/ShapeType.java
@@ -1,8 +1,0 @@
-package br.com.lima.creationalPatterns.factory.halfSimple.enumeration;
-
-public enum ShapeType {
-
-    CIRCLE,
-    RECTANGULE,
-    SQUARE;
-}
