+++ src/main/java/br/com/lima/creationalPatterns/factory/shape/factoryMethod/factory/CircleFactory.java
@@ -1,0 +1,14 @@
+package br.com.lima.creationalPatterns.factory.shape.factoryMethod.factory;
+
+
+import br.com.lima.creationalPatterns.factory.shape.factoryMethod.model.Circle;
+import br.com.lima.creationalPatterns.factory.shape.factoryMethod.model.Shape;
+
+public class CircleFactory extends ShapeFactory{
+
+
+    @Override
+    protected Shape createShape() {
+        return new Circle();
+    }
+}
