+++ src/main/java/br/com/lima/creationalPatterns/factory/shape/factoryMethod/factory/ShapeFactory.java
@@ -1,0 +1,18 @@
+package br.com.lima.creationalPatterns.factory.shape.factoryMethod.factory;
+
+
+import br.com.lima.creationalPatterns.factory.shape.factoryMethod.model.Shape;
+
+public abstract class ShapeFactory {
+
+    public Shape orderShape() {
+        Shape shape = createShape();
+
+        shape.calculateArea();
+
+        return shape;
+    }
+
+    protected abstract Shape createShape();
+
+}
