+++ src/main/java/br/com/lima/creationalPatterns/factory/shape/factoryMethod/model/Shape.java
@@ -1,0 +1,5 @@
+package br.com.lima.creationalPatterns.factory.shape.factoryMethod.model;
+
+public interface Shape {
+    void calculateArea();
+}
