+++ src/main/java/br/com/lima/creationalPatterns/factory/shape/halfSimple/enumeration/ShapeType.java
@@ -1,0 +1,8 @@
+package br.com.lima.creationalPatterns.factory.shape.halfSimple.enumeration;
+
+public enum ShapeType {
+
+    CIRCLE,
+    RECTANGULE,
+    SQUARE;
+}
