+++ src/main/java/br/com/lima/creationalPatterns/factory/shape/halfSimple/model/Shape.java
@@ -1,0 +1,5 @@
+package br.com.lima.creationalPatterns.factory.shape.halfSimple.model;
+
+public interface Shape {
+    void calculateArea();
+}
