+++ src/main/java/br/com/lima/creationalPatterns/factory/shape/simple/enumeration/ShapeType.java
@@ -1,0 +1,8 @@
+package br.com.lima.creationalPatterns.factory.shape.simple.enumeration;
+
+public enum ShapeType {
+
+    CIRCLE,
+    RECTANGULE,
+    SQUARE;
+}
