--- src/main/java/br/com/lima/creationalPatterns/factory/simple/enumeration/ShapeType.java
@@ -1,8 +1,0 @@
-package br.com.lima.creationalPatterns.factory.simple.enumeration;
-
-public enum ShapeType {
-
-    CIRCLE,
-    RECTANGULE,
-    SQUARE;
-}
