--- src/main/java/br/com/lima/creationalPatterns/factory/simple/model/Shape.java
@@ -1,5 +1,0 @@
-package br.com.lima.creationalPatterns.factory.simple.model;
-
-public interface Shape {
-    void calculateArea();
-}
